--- controllers/tiflow_engine_executor_failover.go
@@ -1,1 +1,0 @@
-package controllers
