--- controllers/tiflow_engine_executor_failover_test.go
@@ -1,1 +1,0 @@
-package controllers
