--- controllers/tiflow_engine_executor_member_manager_test.go
@@ -1,1 +1,0 @@
-package controllers
