--- controllers/tiflow_engine_executor_scaler.go
@@ -1,1 +1,0 @@
-package controllers
