--- controllers/tiflow_engine_executor_scaler_test.go
@@ -1,1 +1,0 @@
-package controllers
