--- controllers/tiflow_engine_executor_upgrader.go
@@ -1,1 +1,0 @@
-package controllers
