--- controllers/tiflow_engine_executor_upgrader_test.go
@@ -1,1 +1,0 @@
-package controllers
